+++ src/state.cpp
@@ -583,11 +583,6 @@ Expr State::mkQuoteType(const Expr& t)
   return Expr(mkExprInternal(Kind::QUOTE_TYPE, {t.getValue()}));
 }
 
-Expr State::mkQuoteType(const Expr& t, const Expr& type)
-{
-  return Expr(
-      mkExprInternal(Kind::QUOTE_TYPE, {t.getValue(), type.getValue()}));
-}
 Expr State::mkBuiltinType(Kind k)
 {
   // for now, just use any type