+++ src/type_checker.cpp
@@ -365,26 +365,24 @@ Expr TypeChecker::getTypeAppInternal(std::vector<ExprValue*>& children,
     // which along with how ctypes[i] is the argument itself, has the effect
     // of an implicit upcast.
     bool isQuote = false;
-    bool success = true;
+    bool typeSuccess = true;
     ExprValue* child = children[i];
     if (hdt->getKind() == Kind::QUOTE_TYPE)
     {
-      // if there are two children to Quote, then the second specifies the
-      // expected type. We require that the type matches verbatim.
-      if (hdt->getNumChildren() == 2)
+      // We ensure that the type of the argument is equal to the type of the
+      // quoted term, whose type should both be ground.
+      ExprValue* ct = d_state.lookupType(child);
+      ExprValue* cte = d_state.lookupType(hdt->d_children[0]);
+      Assert (cte->isGround());
+      Assert(ct != nullptr);
+      if (ct != cte)
       {
-        // type should be ground by this point.
-        Assert(hdt->d_children[1]->isGround());
-        ExprValue* ct = d_state.lookupType(child);
-        Assert(ct != nullptr);
-        if (ct != hdt->d_children[1])
-        {
-          success = false;
-          hdt = hdt->d_children[1];
-          child = ct;
-        }
+        typeSuccess = false;
+        hdt = cte;
+        child = ct;
       }
-      if (success)
+      // if the above check was success, we quote
+      if (typeSuccess)
       {
         hdt = hdt->d_children[0];
         isQuote = true;
@@ -395,7 +393,7 @@ Expr TypeChecker::getTypeAppInternal(std::vector<ExprValue*>& children,
       child = d_state.lookupType(child);
       Assert(child != nullptr);
     }
-    if (!success || !match(hdt, child, ctx, visited))
+    if (!typeSuccess || !match(hdt, child, ctx, visited))
     {
       if (out)
       {